ai-model/requirements.txt
langchain_postgres
groq
langchain-groq
langchain_google_genai
langchain
langchain_community
python-dotenv
fastapi
uvicorn
psycopg2-binary
psycopg2
psycopg_binary
psycopg[binary]
sqlalchemy
lxml